[refactoring] popover page
팝오버 페이지 생성
기존 모달형식 페이지보다는
서버 컴포넌트 방식 활용,
그리고 링크 컴포넌트 활용시 페이지
프리렌더링 가능해서 보기좋음
