+++ src/app/(about-layout)/layout.tsx
@@ -3,21 +3,25 @@ import { ReactNode } from "react";
 
 interface AboutLayoutProps {
   children: ReactNode;
+  popover: ReactNode;
 }
 
 export const metadata: Metadata = {
   title: "ABOUT | ou9999.dev",
 };
 
-const AboutLayout = ({ children }: AboutLayoutProps) => {
+const AboutLayout = ({ children, popover }: AboutLayoutProps) => {
   return (
-    <div className="w-dvw h-auto min-h-dvh relative">
-      <div className="w-full flex flex-col justify-center items-center">
-        <div className="mt-5 sm:mt-7 md:mt-10 w-full max-w-276 h-full relative p-3">
-          {children}
+    <>
+      <div className="w-dvw h-auto min-h-dvh relative">
+        <div className="w-full flex flex-col justify-center items-center">
+          <div className="mt-5 sm:mt-7 md:mt-10 w-full max-w-276 h-full relative p-3">
+            {children}
+          </div>
         </div>
       </div>
-    </div>
+      {popover}
+    </>
   );
 };
